+++ mobile/android/android-components/components/browser/state/src/test/java/mozilla/components/browser/state/engine/EngineObserverTest.kt
@@ -110,6 +110,11 @@ class EngineObserverTest {
                 onResult: (Boolean) -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
+            override fun reportBackInStock(
+                url: String,
+                onResult: (String) -> Unit,
+                onException: (Throwable) -> Unit,
+            ) {}
             override fun sendImpressionAttributionEvent(
                 aid: String,
                 onResult: (Boolean) -> Unit,
@@ -226,6 +231,11 @@ class EngineObserverTest {
                 onResult: (Boolean) -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
+            override fun reportBackInStock(
+                url: String,
+                onResult: (String) -> Unit,
+                onException: (Throwable) -> Unit,
+            ) {}
             override fun requestTranslate(
                 fromLanguage: String,
                 toLanguage: String,
@@ -332,6 +342,11 @@ class EngineObserverTest {
             override fun sendImpressionAttributionEvent(
                 aid: String,
                 onResult: (Boolean) -> Unit,
+                onException: (Throwable) -> Unit,
+            ) {}
+            override fun reportBackInStock(
+                url: String,
+                onResult: (String) -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
             override fun requestTranslate(